+++ Finance.js
@@ -1,0 +1,109 @@
+import React, {Component} from 'react';
+/*import MortgageCalculator from "mortgage-calculator-react";
+import CompoundInterest from 'react-compound-interest'; */
+
+export default class Finance extends Component {
+    render() {
+        return (
+            <div>
+              <div class="ui two column padded grid">
+                      <div class="column">
+                      <h1>Financial Aid</h1>
+                     {/*Header */}
+        
+                      {/*HyperLinks */}
+                      <p><a href=" https://www.sfa.ufl.edu/">Office of financial aid</a></p>
+                      <p><a href=" https://umatter.ufl.edu/office/student-financial-affairs/">Financial Aid Hours of Operation</a></p>
+                      <p><a href=" https://admissions.dental.ufl.edu/financial-aid/">Financial Aid Overview</a></p>
+                      <p><a href=" https://www.sfa.ufl.edu/help/">Financial Aid Direct Contact</a></p>
+                            </div>
+                          </div>
+        
+        
+              <div class="ui two column padded grid">
+                      <div class="column">
+                      <h1>Financial Literacy</h1>
+                      {/*Header */}
+        
+                      {/*HyperLinks */}
+                      <p><a href="https://www.usnews.com/education/financial-literacy-college-students">Financial Literacy for College Students</a></p>
+                      <p><a href="https://www.sfa.ufl.edu/resources/">UF Financial Toolkit</a></p>
+                      <p><a href="https://www.sfa.ufl.edu/resources/financial-literacy/">UF Financial Literacy</a></p>
+                      <p><a href="https://learn-and-grow.hr.ufl.edu/courses-registration/financial-literacy-and-retirement-education/">UF Financial Literacy courses</a></p>
+                      <p><a href="https://benefits.hr.ufl.edu/retirement/flare/">UF Financial Literacy course (F.L.A.R.E)</a></p>
+                      <p><a href="https://www.sfa.ufl.edu/news/financial-literacy-workshop/">Financial Literacy Workshop</a></p>
+                      <p><a href="https://distance.ufl.edu/pfft/">UF PFFT</a></p>
+                      <p><a href="https://opportunity.ufl.edu/students/current-students/first-year-students/fall-financial-literacy-workshops/">UF Literacy Workshops</a></p>
+        
+        
+        
+        
+                           
+                           
+                           
+                           
+                           
+                            </div>
+                          </div>
+        
+              <div class="ui two column padded grid">
+                      <div class="column">
+                      <h1>Investing</h1>
+                      {/*Header */}
+        
+                      {/*HyperLinks */}
+                      <p><a href=" https://ufsic.com/index.html">Student Investment Club</a></p>
+                      <p><a href="https://warrington.ufl.edu/gator-student-investment-fund/">Gator Student Investment Fund</a></p>
+                      <p><a href="https://www.investing.com/">General Investment Guide</a></p>
+                      <p><a href="https://www.nerdwallet.com/article/investing/how-to-invest-in-stocks">Investment Starting Guide</a></p>
+                      <p><a href="https://www.investopedia.com/articles/basics/06/invest1000.asp">Stock Investment Basics</a></p>
+                      <p><a href="https://www.nerdwallet.com/article/investing/retirement-investments-beginners-guide">Saving for Retirment </a></p>
+        
+        
+        
+                            </div>
+                          </div>
+                    
+              <div class="ui two column padded grid">
+                      <div class="column">
+                      <h1>Loan Calculators</h1>
+                      {/*Header */}
+        
+                      {/*HyperLinks */}
+                      <p><a href=" https://www.calculator.net/future-value-calculator.html">Future Value Calculator</a></p>
+                      <p><a href=" https://www.bankrate.com/calculators/college-planning/loan-calculator.aspx">Loan Calculator</a></p>
+                      <p><a href=" https://finaid.org/calculators/loanpayments/">Loan Payment Calculator</a></p>
+                      {/*INSTALLED MORTGAGE.JS */}
+        
+                      {/*Put This underneath a dropdown bar, Takes up too much space on the page */}
+
+                       {/*<h2>Mortgage Calculator</h2>
+                       <div>
+                            <MortgageCalculator/>
+                            <MortgageCalculator showPaymentSchedule /> 
+                      </div>
+        
+        <h2>Compound Interest Calculator </h2>
+                      <div>
+                      <CompoundInterest title="Compound Interest Calculator" />
+                       </div>  */}
+        
+        
+        {/*
+        <h2>Interest Calculator</h2>
+                      <script src="../../lib/interest.js"></script>
+                      <script>
+                          var interest_1 = new Interest(1000, 12, 5, {tax:19});
+                      </script> 
+        */}
+                            </div>
+                          </div>      
+                    
+        
+        
+        
+            </div>
+          );
+    }
+
+}
