+++ Health.js
@@ -1,0 +1,108 @@
+import React, {Component} from 'react';
+
+export default class Health extends Component {
+    render() {
+        return (
+            <div className="HealthPage">
+              <div>
+            <div class="header">
+              {/* do not change ^ to header */}
+              <h1>In an emergency call 911. For mental distress call the suicide hotline at 1-800-273-8255</h1>
+            </div>   
+
+            <div class="ui divider"></div>
+      <div class="ui two column padded grid">
+        <div class="column">
+          <h1>STUDENTS</h1>
+
+        <h2>Student Health Care Center</h2> {/*Header */}
+        {/*HyperLinks */}
+        <p><a href=" https://shcc.ufl.edu/">Student Healh Center</a></p>
+
+        <p><a href=" https://shcc.ufl.edu/services/covid19/">Student Healh Center Covid-19</a></p>
+
+        <p><a href=" https://ufhealth.org/uf-student-health-care-center-main-campus">SHCC-UFHealth</a></p>
+
+        <h2>Sexual Health</h2>
+        <p><a href=" https://shcc.ufl.edu/tag/sexual-health/">SHCC-Sexual Health</a></p>
+        <p><a href=" https://gatorwell.ufsa.ufl.edu/health-topic/sexual-health/">Gatorwell- Sexual Health</a></p>
+        <p><a href="  https://www.plannedparenthood.org/health-center/florida/gainesville/32601/gainesville-health-center-2155-90320/std-testing-treatment">Planned Parenthood</a></p>
+        <p><a href="  https://www.cdc.gov/sexualhealth/Default.html">CDC-Sexual Health</a></p>
+        <p><a href="   https://www.pornhub.com/view_video.php?viewkey=ph5f4f9c6bd67e4">Sexual Health-Funny</a></p>
+        {/*For the love of god delete this link or just don't show it during presentation^^^^ */}
+      
+      <p><a href=" https://www.thehotline.org/">National Domestic Violence Hotline</a></p>
+      <p><a href="  https://www.myflfamilies.com/service-programs/abuse-hotline/frequently-asked-questions.shtml">myflfamilies</a></p>
+
+        </div>
+      </div>
+
+        <div class="ui two column padded grid">
+        <div class="column">
+
+          <h1>COVID-19</h1>
+
+          <h2>Covid-19 is a very real and dangerous virus please protect yourself and others</h2>
+          {/* Headers and links for specific websites that deal with Covid-19 */}
+          <h3>CDC - official website</h3>
+          <p><a href=" https://www.cdc.gov/coronavirus/2019-nCoV/index.html">CDC-Covid-19</a></p>
+
+          <h3>CDC informational pamphlet</h3>
+          <p><a href=" https://www.cdc.gov/coronavirus/2019-ncov/communication/print-resources.html?Sort=Date%3A%3Adesc">CDC-print resources</a></p>
+
+          <h3>General Information</h3>
+          <p><a href=" https://www.cdc.gov/coronavirus/2019-ncov/index.html">CDC.gov</a></p>
+
+          <h3>HRSA Covid-19 </h3>
+          <p><a href=" https://www.hrsa.gov/opa/COVID-19-resources">HRSA COVID-19 Resources</a></p>
+
+          <h3>NIH- Covid-19 and substance abuse</h3>
+          <p><a href=" https://www.drugabuse.gov/drug-topics/comorbidity/covid-19-resources">NIH - drug abuse</a></p>
+
+          <h3>NIH - Covid-19</h3>
+          <p><a href=" https://www.nih.gov/coronavirus">NIH - Covid-19</a></p>
+
+          <h3>University of Florida Covid-19 response</h3>
+          <p><a href=" https://coronavirus.ufl.edu/">UF - coronavirus</a></p>
+
+        </div>
+        </div>
+      
+        <div class="ui two column padded grid">
+        <div class="column">
+
+          <h1>MENTAL HEALTH</h1>
+            <h2>Resources</h2>
+          <p><a href=" https://www.mentalhealth.gov/basics/what-is-mental-health">Mental Health.gov/Basics</a></p>
+          <p><a href="   https://www.mentalhealth.gov/">Mental Health.gov</a></p>
+          <p><a href="  https://www.nimh.nih.gov/index.shtml">NIMH.NIH.gov</a></p>
+
+          <h2>Stress Relief</h2>
+          <p><a href=" https://www.youtube.com/watch?v=_f_BnneFanM">Music to relax to </a></p>
+          <p><a href="  https://open.spotify.com/">Spotify </a></p>
+          <p><a href="      https://www.youtube.com/">YouTube </a></p>
+          <p><a href="  https://www.publix.com/pd/publix-chicken-tender-sub/BMO-DSB-100011">Get yourself a chicken tender sub </a></p>
+          {/*Code a submit link button here  */}
+        
+
+        </div>
+        </div>
+
+        <div class="ui two column padded grid">
+        <div class="column">
+          <h1>Shands</h1>
+          <p><a href="https://ufhealth.org/">ufhealth.org</a></p>
+          <p><a href=" https://ufhealth.org/shands-university-florida">ufhealth-shands</a></p>
+          <p><a href="https://ufhealth.org/uf-health-shands-hospital-patient-guide/telephone-directory">Shands Patient Directory</a></p>
+        
+        </div>
+        </div>
+            </div>
+              {/* INSERT SUBMIT BUTTON THAT SENDS AN EMAIL OR DATA OR WHATEVER TO SUBMIT A NEW LINK */}
+          
+            </div>
+
+          );
+    }
+
+}
