+++ Home.js
@@ -1,0 +1,11 @@
+import React, {Component} from 'react';
+
+export default class Home extends Component {
+    render() {
+        return (
+            <div>
+              <h1>Home</h1>
+            </div>
+          );
+    }
+}
